Sidebar redesign, related articles, sharper corners, lighter chrome.
