+++ src/components/NavigationDocs.jsx
@@ -3,22 +3,8 @@ import clsx from 'clsx'
 import Link from 'next/link'
 import { AnimatePresence, LayoutGroup, motion, useReducedMotion } from 'framer-motion'
 import { useEffect, useState } from 'react'
-import {
-  Home,
-  Users,
-  Handshake,
-  FileText,
-  Code,
-  Monitor,
-  Wrench,
-  ChevronRight,
-  BookOpen,
-  FolderOpen,
-  Tag,
-  Rss,
-  Github,
-  Info
-} from 'lucide-react'
+import { ChevronRight, Rss, Github, Info } from 'lucide-react'
+import { docsNavigation } from '@/data/docsNavigation'
 
 function NavLink({ href, active, children }) {
     const reduceMotion = useReducedMotion()
@@ -62,72 +48,6 @@ function TagChip({ href, active, children }) {
         </Link>
     )
 }
-
-export const docsNavigation = [
-    {
-        title: 'PAGES',
-        icon: <Home className="h-3.5 w-3.5" />,
-        defaultOpen: true,
-        pages: [
-            { title: 'Introduction', href: '/' },
-            { title: 'All Content', href: '/content' },
-            { title: 'Partners', href: '/partner' },
-            { title: 'Team', href: '/team' },
-        ],
-    },
-    {
-        title: 'CATEGORIES',
-        icon: <FolderOpen className="h-3.5 w-3.5" />,
-        defaultOpen: true,
-        pages: [
-            { title: 'Guides', href: '/categories/guides' },
-            { title: 'Software', href: '/categories/software' },
-            { title: 'Benchmarking', href: '/categories/benchmarking' },
-            { title: 'Essay', href: '/categories/essay' },
-            { title: 'Hardware', href: '/categories/hardware' },
-        ],
-    },
-    {
-        title: 'TAGS',
-        icon: <Tag className="h-3.5 w-3.5" />,
-        defaultOpen: false,
-        display: 'chips',
-        pages: [
-            { title: 'AI', href: '/tags/ai' },
-            { title: 'Apps', href: '/tags/apps' },
-            { title: 'Arch', href: '/tags/arch' },
-            { title: 'Archive', href: '/tags/archive' },
-            { title: 'Benchmarking', href: '/tags/benchmarking' },
-            { title: 'COSMIC', href: '/tags/cosmic' },
-            { title: 'Customization', href: '/tags/customization' },
-            { title: 'DaVinci Resolve', href: '/tags/davinci-resolve' },
-            { title: 'Desktop Environments', href: '/tags/desktop-environments' },
-            { title: 'Distros', href: '/tags/distros' },
-            { title: 'Docker', href: '/tags/docker' },
-            { title: 'Essay', href: '/tags/essay' },
-            { title: 'Fedora', href: '/tags/fedora' },
-            { title: 'Git', href: '/tags/git' },
-            { title: 'Gnome', href: '/tags/gnome' },
-            { title: 'Guides', href: '/tags/guides' },
-            { title: 'Hardware', href: '/tags/hardware' },
-            { title: 'Homelab', href: '/tags/homelab' },
-            { title: 'KDE', href: '/tags/kde' },
-            { title: 'Khadas', href: '/tags/khadas' },
-            { title: 'Linux', href: '/tags/linux' },
-            { title: 'Manjaro', href: '/tags/manjaro' },
-            { title: 'Mobile', href: '/tags/mobile' },
-            { title: 'News', href: '/tags/news' },
-            { title: 'OpenSUSE', href: '/tags/opensuse' },
-            { title: 'Pine64', href: '/tags/pine64' },
-            { title: 'Proxmox', href: '/tags/proxmox' },
-            { title: 'RaspberryPi', href: '/tags/raspberrypi' },
-            { title: 'Red Hat', href: '/tags/red-hat' },
-            { title: 'Self-Host', href: '/tags/self-host' },
-            { title: 'Single Board Computers', href: '/tags/single-board-computers' },
-            { title: 'Windows', href: '/tags/windows' },
-        ],
-    },
-]
 
 export function NavigationDocs({ className }) {
     return (
